requirements.txt
pyyaml
aws-cdk.core==1.*
aws-cdk.aws-ecr==1.*
aws-cdk.aws-iam==1.*
aws-cdk.pipelines==1.*